[Feat] 비동기 로딩 에 사용될 캔버스 프리팹 추가 및 재생
