--- Assets/KYH_card/Card Script.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6a025938160697e458bf3a65ac50c653
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
