--- Assets/KYH_card/Card Script/CardAnimator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6c3af20f40d30e84786a4ac2d1e854ad
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
