--- Assets/KYH_card/Card Script/CardEffect.cs
@@ -1,8 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public abstract class CardEffect : MonoBehaviour
-{
-    public abstract void ApplyEffect(GameObject player);
-}
