--- Assets/KYH_card/Card Script/CardEffect.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b979b2255b90714084eea726286428b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
