--- Assets/KYH_card/Card Script/CardManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6e3563f4612c3694683e6b8072d6285d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
