+++ Assets/KYH_card/Game Scene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: aa589861cb717354b82a5dba0439d6e0
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
