+++ Assets/KYH_card/SceneLoadingManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 47b5c85b7bd38dc419ef08997f00214c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
