+++ Assets/KYH_card/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aed7cdff00179a241bc9b2a655417266
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
