+++ Assets/KYH_card/Scripts/Card Script/CardDataSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d373a57fa6e55da4da12af025233a94a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
