+++ Assets/KYH_card/Scripts/Card Script/CardEffect.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public abstract class CardEffect : MonoBehaviour
+{
+    public abstract void ApplyEffect(GameObject player);
+}
