+++ Assets/KYH_card/Scripts/Card Script/CardUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8400944fe8402bb49a0b4eb964f647c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
