+++ Assets/KYH_card/Scripts/Card Script/FlipCard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3e26e5e2beff5754da6e6aaf416ea1de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
