+++ Assets/KYH_card/Scripts/Scene scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0fdb6c4891dcfe14083d0bef296c4122
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
