+++ Assets/KYH_card/Scripts/Scene scripts/Game End Manager.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class GameEndManager : MonoBehaviour
+{
+    public void Update()
+    {
+        if(Input.GetKeyDown(KeyCode.Space))
+        {
+            SceneLoadingManager.Instance.LoadSceneAsync("CardTest");
+        }
+    }
+}
