+++ Assets/KYH_card/Scripts/Scene scripts/Game End Manager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f082e5c855f19a43ac4fa8b17cb4779
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
