+++ Assets/KYH_card/Scripts/Scene scripts/LoadingUIManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a49af04dd4ebd34398b73a95232cc4e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
