+++ Assets/KYH_card/Scripts/Scene scripts/SceneLoadingManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a07e3487eef8341459b96c534adaa671
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
