+++ ProjectSettings/EditorBuildSettings.asset
@@ -8,4 +8,10 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/JGH/Scene/Test5.unity
     guid: 872f9cd8fe02f4af5901d4b92a381c09
+  - enabled: 1
+    path: Assets/KYH_card/CardTest.unity
+    guid: 6a2a6c4eff1630742a44b03081656f88
+  - enabled: 1
+    path: Assets/KYH_card/Game Scene.unity
+    guid: aa589861cb717354b82a5dba0439d6e0
   m_configObjects: {}
